% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{add}
\alias{add}
\title{Add two numbers.

This is the description section (one paragraph)

This is the details section (zero or more paragraphs)}
\usage{
add(a, b)
}
\arguments{
\item{a}{First number to add}

\item{b}{Second number to add}
}
\value{
The added numbers
}
\description{
Add two numbers.

This is the description section (one paragraph)

This is the details section (zero or more paragraphs)
}
\examples{
add(1, 2)

}
