% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{calc_age_specific_fertility_rate}
\alias{calc_age_specific_fertility_rate}
\title{Calculate ASFR from TFR and fertility distribution.}
\usage{
calc_age_specific_fertility_rate(fertility_rate, births_proportion)
}
\arguments{
\item{fertility_rate}{vector of annual fertility rate values.}

\item{births_proportion}{array of proportion of births by 5 year age group
15-49.}
}
\value{
array of age-specific fertility rate by single-year of age 15-49.
}
\description{
Splits each 5 year age group into 5, distributing proportion of births (asfd)
evenly between the single ages. Then multiply each age group by the
corresponding fertility rate for that that age group and year.
}
\keyword{internal}
