% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{calc_net_migration}
\alias{calc_net_migration}
\title{Calculate net migration for single years using beers coefficients.}
\usage{
calc_net_migration(total_net_migr, net_migr_age_dist)
}
\arguments{
\item{total_net_migr}{The total net migration by 5 year age groups.}

\item{net_migr_age_dist}{The net migration age distribution.}
}
\value{
Single year migration data.
}
\description{
This uses beers coefficients to calculate net migration for a single age
from data aggregated into 5 year age groups.
}
\keyword{internal}
