% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{convert_type}
\alias{convert_type}
\title{Convert filtered data to correct type.}
\usage{
convert_type(data)
}
\arguments{
\item{data}{}
}
\value{
The converted data
}
\description{
Now data has been filtered to only data for a specific property we can
ensure that the data type is set to numeric where possible.
}
\keyword{internal}
