% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_age_14_hiv_population}
\alias{dimensions_age_14_hiv_population}
\title{Get dimensions for age 14 hiv population tag.}
\usage{
dimensions_age_14_hiv_population(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
ART stage, CD4 category, sex and year.
}
\keyword{internal}
