% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_age_sex_year}
\alias{dimensions_age_sex_year}
\title{Get single age between 0 and 80, sex and year dimensions.}
\usage{
dimensions_age_sex_year(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get single age between 0 and 80, sex and year dimensions.
}
\keyword{internal}
