% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_agegr_sex_year}
\alias{dimensions_agegr_sex_year}
\title{Get age group, sex and year dimensions.}
\usage{
dimensions_agegr_sex_year(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get age group, sex and year dimensions.
}
\keyword{internal}
