% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_agegr_year_alternative}
\alias{dimensions_agegr_year_alternative}
\title{Get age groups between 15 and 49 and year dimensions alternatives.}
\usage{
dimensions_agegr_year_alternative(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Used for different Spectrum version than `dimensions_agegr_year`.
}
\keyword{internal}
