% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_art}
\alias{dimensions_art}
\title{Get artdur and year dimensions.}
\usage{
dimensions_art(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get artdur and year dimensions.
}
\keyword{internal}
