% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_cd4}
\alias{dimensions_cd4}
\title{Get CD4 stage and age category dimensions.}
\usage{
dimensions_cd4(proj_years = NULL)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get CD4 stage and age category dimensions.
}
\keyword{internal}
