% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_dimensions.R
\name{dimensions_cd4_progress}
\alias{dimensions_cd4_progress}
\title{Get dimensions for adult CD4 progress data.}
\usage{
dimensions_cd4_progress(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Stages of infection and age categories.
}
\keyword{internal}
