% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_hiv_population}
\alias{extract_hiv_population}
\title{Extract people living with HIV (PLHIV) from PJNZ file}
\usage{
extract_hiv_population(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to PJNZ file to extract the data from.}
}
\value{
3D array of HIV population data separated by year, sex and age.
}
\description{
Get HIV population data separated by year, sex and age. Final age is open
ended age group 80+.
}
\examples{
pjnz_path <- system.file("testdata", "Botswana2017.PJNZ", package = "specio")
hiv_population <- extract_hiv_population(pjnz_path)

}
