% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epp_results.R
\name{extract_incidence_prevalence}
\alias{extract_incidence_prevalence}
\title{Extract incidence and prevalence for single region.}
\usage{
extract_incidence_prevalence(break_index, spt_data, no_of_years)
}
\arguments{
\item{break_index}{The index of the line break in the spt_data where data
about this region ends.}

\item{spt_data}{The complete spt data read from a file.}

\item{no_of_years}{The number of years in the data set, equivalent to the
number of rows we return.}
}
\value{
A data frame, one row for each year showing the HIV incidence,
prevalence and population (if available).
}
\description{
Gets the incidence, prevalence and population data (if available) for a
single region, e.g. National, Urban or Rural.
}
\keyword{internal}
