% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_agegr_and_sex_dimnames}
\alias{get_agegr_and_sex_dimnames}
\title{Get the dimensions and names for agegr, sex and proj years data.}
\usage{
get_agegr_and_sex_dimnames(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get the dimensions and names for agegr, sex and proj years data.
}
\keyword{internal}
