% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_agegr_labels}
\alias{get_agegr_labels}
\title{Get the age group labels.}
\usage{
get_agegr_labels()
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get the age group labels.
}
\keyword{internal}
