% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_array_data}
\alias{get_array_data}
\title{Get the notes column for a particular property from DP data.}
\usage{
get_array_data(tag, tag_data, rows, cols)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Filtered tag data.}

\item{tag_data}{Tag data to be parsed.}

\item{...}{Other params made available calling function.}
}
\description{
Get the notes column for a particular property from DP data.
}
\keyword{internal}
