% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{get_array_property_data}
\alias{get_array_property_data}
\title{Get data for an array property from the xml_node.}
\usage{
get_array_property_data(node)
}
\arguments{
\item{node}{The xml node.}
}
\value{
Parsed property data.
}
\description{
Will get array data from one xml_node. If node represents a vector then a
vector is returned. If node repsents a matrix return matrix.
}
\keyword{internal}
