% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{get_art_mortality}
\alias{get_art_mortality}
\title{Retrieve and tidy ART mortality data.}
\usage{
get_art_mortality(dp_data)
}
\arguments{
\item{dp_data}{The full set of dp data to extract data from.}
}
\description{
Retrieve and tidy ART mortality data.
}
\keyword{internal}
