% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_art_mortality_rates}
\alias{get_art_mortality_rates}
\title{Get ART mortality rate data from DP data.}
\usage{
get_art_mortality_rates(tag, tag_data, metadata, proj_years)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{metadata}{Metadata params made available by tag configuration. Must
contain rows.}

\item{proj_years}{Years the projection is for.}
}
\description{
Requires some special handling as we need the data separated by
time on ART but this isn't always provided by Spectrum.
}
\keyword{internal}
