% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_art_pop}
\alias{get_art_pop}
\title{Extract ART population data from PJNZ file}
\usage{
get_art_pop(pjnz_path, long_format = FALSE)
}
\arguments{
\item{pjnz_path}{Path to PJNZ file to extract the data from.}

\item{long_format}{If TRUE then data is returned in long format.}
}
\value{
3D array of ART population data separated by year, sex and age.
}
\description{
Get ART population data separated by year, sex and age. Final age is open
ended age group 80+.
}
\examples{
pjnz_path <- system.file("testdata", "Botswana2017.PJNZ", package = "specio")
art_population <- get_art_pop(pjnz_path)
art_pop_long <- get_art_pop(pjnz_path, TRUE)

}
