% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_cd4_array_data}
\alias{get_cd4_array_data}
\title{Get array data for a cd4 property from DP data.}
\usage{
get_cd4_array_data(tag, tag_data, metadata, proj_years)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{metadata}{Metadata params made available by tag configuration. Must
contain rows, and a func to define array dimensions and names.}

\item{proj_years}{Years the projection is for.}
}
\description{
Require some special handling because each row contains data for an age
group and a cd4 stage.
}
\keyword{internal}
