% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{get_census_data}
\alias{get_census_data}
\title{Extract and prepare census data from ProjectionSet.}
\usage{
get_census_data(projection_set, input_mode)
}
\arguments{
\item{projection_set}{An xml_nodeset representing a ProjectionSet java
object.}
}
\description{
Helper function which handles extracting and tidying census data from the
xml_nodeset representing ProjectionSet java object.
}
\keyword{internal}
