% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{get_concentrated_epidemic_data}
\alias{get_concentrated_epidemic_data}
\title{Get concentrated epidemic data from EPP data set.}
\usage{
get_concentrated_epidemic_data(epp_set)
}
\arguments{
\item{epp_set}{xmlnodeset containing concetrated epidemic data.}
}
\value{
List containing concentrated epidemic data.
}
\description{
Gets all data related to a concentrated epidemic from EPP nodeset.
}
\keyword{internal}
