% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_data}
\alias{get_data}
\title{Get the data from the 'Data' column and onwards if present.}
\usage{
get_data(dp_data)
}
\arguments{
\item{dp_data}{The subsetted dp data to extract data from.}
}
\description{
Gets the data from the 'Data' column and onwards. If no data is found
returns null. Also tries to set rownames as the 'Notes' where these are
present. This expects to receive only the relevant rows, this does the work
to filter the relevant columns only.
}
\keyword{internal}
