% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_data_from_pjnz}
\alias{get_data_from_pjnz}
\title{Get data from PJNZ file for a single property}
\usage{
get_data_from_pjnz(pjnz_path, property, long_format, long_format_label)
}
\arguments{
\item{pjnz_path}{Path to PJNZ file to get data from.}

\item{property}{The name of the property to return data for.}

\item{long_format}{If TRUE data is converted to a long format.}

\item{long_format_label}{If long_format is TRUE then this is used as column
name for the long format data.}
}
\value{
Data extracted from PJNZ.
}
\description{
Extract the data for a single property from the PJNZ file. This accesses the
data from the DP file contained within the PJNZ zip.
}
\keyword{internal}
