% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_data_indices}
\alias{get_data_indices}
\title{Get the start and end indices of a particular property within full dataset.}
\usage{
get_data_indices(property, dp_data)
}
\arguments{
\item{property}{Property to get.}

\item{dp_data}{Full set of untidied dp data.}
}
\value{
List contianing the start and end indices of the data.
}
\description{
End index is inclusive of row containing <End> tag.
}
\keyword{internal}
