% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_default_array}
\alias{get_default_array}
\title{Get default array data}
\usage{
get_default_array(metadata, proj_years)
}
\arguments{
\item{metadata}{List containing value for array and dimensions of array.}

\item{proj_years}{Years the projection is for.}
}
\description{
Defaults to creating a data frame of a single value of dimensions specified
by metadata.
}
\keyword{internal}
