% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_default_art_mortality_rates}
\alias{get_default_art_mortality_rates}
\title{Get default ART mortality rates.}
\usage{
get_default_art_mortality_rates(proj_years)
}
\arguments{
\item{proj_years}{Years the projection is for.}
}
\description{
Defaults to using 1.0 for each point.
}
\keyword{internal}
