% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_description_indices}
\alias{get_description_indices}
\title{Get the indices for which there is a non empty field in the description col.}
\usage{
get_description_indices(start_index, end_index, dp_data)
}
\arguments{
\item{start_index}{Index to start looking from.}

\item{end_index}{Index to look up to.}

\item{dp_data}{The full set of dp data.}
}
\description{
Gets all indices of not empty description fields, excluding <Value> tag.
Use these to identify separate blocks of data within the dp data. Get these
between two indices
}
\keyword{internal}
