% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_dp_data}
\alias{get_dp_data}
\title{Get data from DP file.}
\usage{
get_dp_data(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to zip containing DP file.}
}
\value{
The data from file read as a csv.
}
\description{
Get data from DP file.
}
