% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_dp_notes}
\alias{get_dp_notes}
\title{Get the notes column for a particular property from DP data.}
\usage{
get_dp_notes(property, dp_data)
}
\arguments{
\item{property}{Property to get notes for.}

\item{dp_data}{Full set of untidied dp data.}
}
\value{
The notes.
}
\description{
Get the notes column for a particular property from DP data.
}
\keyword{internal}
