% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_dp_property}
\alias{get_dp_property}
\title{Get a property from the DP file.}
\usage{
get_dp_property(property, dp_data)
}
\arguments{
\item{property}{Property to get.}

\item{dp_data}{Full set of untidied dp data.}
}
\value{
The tidied property data.
}
\description{
Finds a property via a matching "Tag" and extracts the data contained in
the tag.
}
\keyword{internal}
