% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tags.R
\name{get_dp_tags}
\alias{get_dp_tags}
\title{Get tags for a specific Spectrum version.}
\usage{
get_dp_tags(version)
}
\arguments{
\item{version}{Spectrum version to get tags for.}
}
\value{
List of properties mapped to their tag in Spectrum DP file.
}
\description{
Get tags for a specific Spectrum version.
}
\keyword{internal}
