% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_eligibility_pop_data}
\alias{get_eligibility_pop_data}
\title{Get ART eligibility population data.}
\usage{
get_eligibility_pop_data(tag, tag_data, metadata, proj_years)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{metadata}{Metadata params made available by tag configuration.}

\item{agegr_labels}{Age group labels.}
}
\description{
Get ART eligibility population data.
}
\keyword{internal}
