% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{get_enum_property}
\alias{get_enum_property}
\title{Get character value of enum property from the xml_node.}
\usage{
get_enum_property(node)
}
\arguments{
\item{node}{The xml node.}
}
\value{
Parsed property data.
}
\description{
Get the enum property of the xml node as a character.
}
\keyword{internal}
