% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_eppxml_workset}
\alias{get_eppxml_workset}
\title{Get EPP workset from xml file.}
\usage{
get_eppxml_workset(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to zip containing workset xml.}
}
\value{
The workset as a list.
}
\description{
Gets all properties of the EPP workset from xml file inside zip at specified
path.
}
