% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_fert_rat}
\alias{get_fert_rat}
\title{Get fertility ratio property}
\usage{
get_fert_rat(tag, dp_data)
}
\value{
The tidied property data.
}
\description{
Get fertility ratio property
}
\keyword{internal}
