% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{get_field_data}
\alias{get_field_data}
\title{Get field data from xml node representing serialised field.}
\usage{
get_field_data(field_node)
}
\arguments{
\item{field_node}{Node representing a serialised field.}
}
\description{
Function relies on knowing the structure in which fields are persisted in
the xml. e.g.
}
\details{
<void class=".." method="getField">
  <string>fieldName</string>
  <void method="set">
    <object idref="../>
    <double>fieldValue</double>
  </void>
</void>
}
\keyword{internal}
