% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{get_fields}
\alias{get_fields}
\title{Get field data from parent node.}
\usage{
get_fields(node)
}
\arguments{
\item{node}{The node containing field data.}
}
\description{
Get all key value pair field data containing in within the node.
}
\keyword{internal}
