% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{get_filename_from_extension}
\alias{get_filename_from_extension}
\title{Get filename from extension.}
\usage{
get_filename_from_extension(file_extension, path_to_zip)
}
\arguments{
\item{file_extension}{The file extension to look for.}

\item{path_to_zip}{The path to the zip file to look in.}
}
\value{
The name of the located file.
}
\description{
Helper method to get the full filename within a zip matching extension. This
expects there to be only one file in the zip with the extension.
}
\keyword{internal}
