% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{get_hh_survey_data}
\alias{get_hh_survey_data}
\title{Extract HH survey data from ProjectionSet.}
\usage{
get_hh_survey_data(projection_set)
}
\arguments{
\item{projection_set}{An xml_nodeset representing a ProjectionSet java
object.}
}
\description{
Helper function which handles extracting ANC data from the xml_nodeset
representing ProjectionSet java object.
}
\keyword{internal}
