% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{get_impact_model_params}
\alias{get_impact_model_params}
\title{Extract AIM module parameters.}
\usage{
get_impact_model_params(dp_data, proj_years)
}
\arguments{
\item{dp_data}{The full set of dp data to extract data from.}

\item{proj_years}{Active years for the projection.}
}
\value{
List of impact model params.
}
\description{
Extract AIM module parameters.
}
\keyword{internal}
