% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epp_results.R
\name{get_incid_prev_results}
\alias{get_incid_prev_results}
\title{Get incidence or prevalence as a matrix from the list of matrices
representing the uncertainty results data.}
\usage{
get_incid_prev_results(results, property)
}
\arguments{
\item{property}{The property to get from the series data. Either "incid" for
incidence or "prev" for prevalence.}

\item{series_data}{The extracted series data as a list of matrices, one list
entry for each resample.}
}
\description{
Get incidence or prevalence as a matrix from the list of matrices
representing the uncertainty results data.
}
\keyword{internal}
