% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjn_metadata.R
\name{get_pjn_country_metadata}
\alias{get_pjn_country_metadata}
\title{Get country name and code from PJNZ}
\usage{
get_pjn_country_metadata(pjnz)
}
\arguments{
\item{pjnz}{Path to PJNZ file.}
}
\value{
The country metadata.
}
\description{
Get country name and code from PJNZ
}
