% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_pjn_data}
\alias{get_pjn_data}
\title{Get data from PJN file}
\usage{
get_pjn_data(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to PJNZ zip file containing PJN file.}
}
\value{
The data from file read as a csv.
}
\description{
Get data from PJN file
}
\keyword{internal}
