% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjn_metadata.R
\name{get_pjn_metadata}
\alias{get_pjn_metadata}
\title{Get projection metadata from PJN file data}
\usage{
get_pjn_metadata(pjn_data)
}
\arguments{
\item{pjn_data}{Data frame of PJN file data}
}
\description{
Get projection metadata from PJN file data
}
\keyword{internal}
