% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjn_metadata.R
\name{get_pjn_region_code}
\alias{get_pjn_region_code}
\title{Get subnational region code from parsed PJN}
\usage{
get_pjn_region_code(pjn)
}
\description{
Get subnational region code from parsed PJN
}
\keyword{internal}
