% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_pjnz_csv_data}
\alias{get_pjnz_csv_data}
\title{Get csv data from file containing within PJNZ}
\usage{
get_pjnz_csv_data(pjnz_path, extension)
}
\arguments{
\item{pjnz_path}{Path to PJNZ zip file.}

\item{extension}{The file extension to look for.}
}
\value{
The data from file read as a csv.
}
\description{
Get csv data from file containing within PJNZ
}
\examples{
pjnz <- system.file("testdata", "Botswana2018.PJNZ", package = "specio")
specio:::get_pjnz_csv_data(pjnz, "DP")
specio:::get_pjnz_csv_data(pjnz, "PJN")
}
\keyword{internal}
