% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{get_projection_years}
\alias{get_projection_years}
\title{Get years of the projection from dp data}
\usage{
get_projection_years(dp_data)
}
\arguments{
\item{dp_data}{The dp data to get the projections years from.}
}
\value{
Sequence of projection years.
}
\description{
Get years of the projection from dp data
}
\keyword{internal}
