% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{get_property}
\alias{get_property}
\title{Get data for a property in a nodeset.}
\usage{
get_property(nodeset, property)
}
\arguments{
\item{nodeset}{The nodeset to get the property from.}

\item{property}{The property to get.}
}
\value{
Parsed property data.
}
\description{
Locates an xml property wthin the nodeset, parses the value depending on
the type of the property and returns the parsed value.
}
\keyword{internal}
