% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{get_property_data}
\alias{get_property_data}
\title{Get data for a property from full dp_data via its tag.}
\usage{
get_property_data(property, dp_data)
}
\arguments{
\item{property}{The property to get data for.}

\item{dp_data}{The full dp dataset.}
}
\value{
The parsed data
}
\description{
Gets the list of tags for a particular property, identifies which one
should be used and then invokes the configured function for that tag.
}
\keyword{internal}
