% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tags.R
\name{get_property_tags}
\alias{get_property_tags}
\title{Get tags mapping for a property.}
\usage{
get_property_tags(property)
}
\arguments{
\item{property}{Property to get the tags for.}
}
\value{
List of possible tags used to refer to the property, in order of
which they should be used. Also returns the function which should be used
to parse the data for the property from the full set of DP data.
}
\description{
This gets list of tags which may be used to store data for the property
within the DP file. Also specifies for each of these properties the function
which should be used to extract the data from full set of DP data.
}
\keyword{internal}
