% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_raw_tag_data}
\alias{get_raw_tag_data}
\title{Get the raw data for a particular property within full dataset.}
\usage{
get_raw_tag_data(tag, dp_data)
}
\arguments{
\item{dp_data}{Full set of untidied dp data.}

\item{property}{Property to get.}
}
\value{
The subsetted data.
}
\description{
This just finds the start an end index of the tag representing the property
within the full dataset and returns the subset.
}
\keyword{internal}
