% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epp_results.R
\name{get_resample_count}
\alias{get_resample_count}
\title{Pull out the count from a string e.g. "COUNT 1.0"}
\usage{
get_resample_count(count)
}
\description{
Pull out the count from a string e.g. "COUNT 1.0"
}
\keyword{internal}
