% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{get_scale_cd4_mortality}
\alias{get_scale_cd4_mortality}
\title{Interpret CD4 mortality scale from version number.}
\usage{
get_scale_cd4_mortality(valid_version)
}
\arguments{
\item{valid_version}{The version number.}
}
\description{
Interpret CD4 mortality scale from version number.
}
\keyword{internal}
