% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_specpop_dimnames}
\alias{get_specpop_dimnames}
\title{Get the dimensions and names for specpop data.}
\usage{
get_specpop_dimnames(proj_years)
}
\arguments{
\item{proj_years}{Vector of years the projection is for.}
}
\description{
Get the dimensions and names for specpop data.
}
\keyword{internal}
