% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{get_spectrum_version}
\alias{get_spectrum_version}
\title{Get the version of Spectrum the dp data is for.}
\usage{
get_spectrum_version(dp_data)
}
\arguments{
\item{dp_data}{The dp_data to get the version for.}
}
\description{
Find the version of the Spectrum which created the dp data by inspecting a
tag.
}
\keyword{internal}
