% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_tag_notes}
\alias{get_tag_notes}
\title{Get the notes column for a particular property from DP data.}
\usage{
get_tag_notes(tag_data)
}
\arguments{
\item{tag_data}{Filteres tag data.}
}
\value{
The notes.
}
\description{
Get the notes column for a particular property from DP data.
}
\keyword{internal}
