% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_total_population}
\alias{get_total_population}
\title{Get total population property.}
\usage{
get_total_population(tag, tag_data, metadata, proj_years)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{metadata}{Metadata params made available by tag configuration.}

\item{...}{Other params made available calling function.}
}
\description{
Get total population property.
}
\keyword{internal}
