% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_vector_data}
\alias{get_vector_data}
\title{Get vector data for a particular property from DP data.}
\usage{
get_vector_data(tag, tag_data, metadata, proj_years)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{metadata}{Metadata params made available by tag configuration. Must
contain cols to get the data for.}

\item{proj_years}{Years the projection is for.}
}
\description{
Get vector data for a particular property from DP data.
}
\keyword{internal}
