% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_women_on_art}
\alias{get_women_on_art}
\title{Get women on ART data.}
\usage{
get_women_on_art(tag, tag_data, metadata, agegr_labels)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{metadata}{Metadata params made available by tag configuration.}

\item{agegr_labels}{Age group labels.}
}
\description{
Handles case when the data only has a single value for women on ART. We
need 1 data point for each age group so just repeat the single value.
}
\keyword{internal}
