% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{next_index}
\alias{next_index}
\title{Get the next index of a tag.}
\usage{
next_index(tag, start_index, column, dp_data)
}
\arguments{
\item{tag}{Tag to look for}

\item{start_index}{Index to start looking from.}

\item{column}{Column to look for the tag in.}

\item{dp_data}{The dp data to look in.}
}
\description{
Get the row index of the next occurance of a tag in a specified column
starting from a specified row.
}
\keyword{internal}
