% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{parse_array}
\alias{parse_array}
\title{Parse xml node containing array to vector.}
\usage{
parse_array(xml_node)
}
\arguments{
\item{xml_node}{`xml_node` object representing an array.}
}
\value{
A vector representing the array.
}
\description{
Parses XML representation of java array from EPP .xml file.
}
\details{
Note that the Java XML representation for an array can have omitted array
entries. This indicate that that value should be the default value for the
type of the array e.g. can get an array of length 5 with only 3 elements in
it. For the missing elements we should use default values at their index in
the array.

The default values are
Java type | R type    | Default value
int       | integer   | 0
double    | numeric   | 0
boolean   | logical   | FALSE
String    | character | ""
}
\keyword{internal}
