% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{parse_matrix}
\alias{parse_matrix}
\title{Parse xml node representing matrix to an R matrix.}
\usage{
parse_matrix(xml_node)
}
\arguments{
\item{xml_node}{`xml_node` object representing a matrix.}
}
\value{
Matrix
}
\description{
Parses XML representation of java array of arrays from EPP .xml file.
}
\keyword{internal}
