% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{parse_region_data}
\alias{parse_region_data}
\title{Tidy subpopulation data about a specific region.}
\usage{
parse_region_data(region_data, start_year, end_year, epi_type)
}
\arguments{
\item{region_data}{Data about a single region for a specific subpop.}

\item{start_year}{Start year of data set.}

\item{end_year}{End year of the data set.}

\item{epidemic_type}{Type of the epidemic. Concentrated or generalized.}
}
\description{
Tidy subpopulation data about a specific region.
}
\keyword{internal}
