% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{parse_survey}
\alias{parse_survey}
\title{Parse survey data into expected return format.}
\usage{
parse_survey(survey)
}
\arguments{
\item{nodeset}{for a survey}
}
\value{
parsed survey data
}
\description{
Parses survey data as stored in 2019 version of spectrum.
}
\keyword{internal}
