% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{read_epp_data}
\alias{read_epp_data}
\title{Read EPP fitting surveillance data}
\usage{
read_epp_data(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to Spectrum PJNZ file.}
}
\value{
List of data frames containing prevalence and incidence data.
}
\description{
Reads the HIV prevalence from sentinel surveillance and household
surveillance data from the EPP .xml file within a PJNZ.
}
\details{
EPP projection sets are identified in the .xml file by searching the XML tree
for tag "object", and then selecting objects with "class" attribute equal to
"epp2011.core.sets.ProjectionSet".
}
\examples{
pjnz_path <- system.file("testdata", "Botswana2018.PJNZ", package="specio")
read_epp_data(pjnz_path)

}
