% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hivproj_param.R
\name{read_hivproj_param}
\alias{read_hivproj_param}
\title{Read params from DP and PJN file required for HIV model.}
\usage{
read_hivproj_param(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to PJNZ zip file containing DP file and PJN file.}
}
\value{
List of model params.
}
\description{
Read params from DP and PJN file required for HIV model.
}
\examples{
pjnz_path <- system.file("testdata", "Botswana2017.PJNZ", package = "specio")
read_hivproj_param(pjnz_path)
}
