% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjn_metadata.R
\name{read_pjn_metadata}
\alias{read_pjn_metadata}
\title{Get country name, iso3 code and spectrum region code from PJNZ}
\usage{
read_pjn_metadata(pjnz)
}
\arguments{
\item{pjnz}{Path to PJNZ file.}
}
\value{
The country and region code metadata
}
\description{
Get country name, iso3 code and spectrum region code from PJNZ
}
