% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pjnz_metadata.R
\name{read_pjnz_metadata}
\alias{read_pjnz_metadata}
\title{Extract projection metadata from PJNZ}
\usage{
read_pjnz_metadata(pjnz_path)
}
\arguments{
\item{pjnz_path}{path to PJNZ file}
}
\value{
List consisting of:
  * country name,
  * Country ISO Alpha-3 code,
  * Country ISO numeric code (3-digit)
  * Spectrum subnational region
  * Spectrum subnational region code
  * Spectrum version
  * Projection name
}
\description{
Extract projection metadata from PJNZ
}
\details{
Returns metadata stored in the .PJN file with .PJNZ archive.
}
