% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjn_metadata.R
\name{read_spectrum_region_code}
\alias{read_spectrum_region_code}
\title{Read spectrum region code from PJN file in PJNZ}
\usage{
read_spectrum_region_code(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to PJNZ file to extract the data from.}
}
\value{
Spectrum region code
}
\description{
Read spectrum region code from PJN file in PJNZ
}
\examples{
pjnz_path <- system.file("testdata", "Botswana2017.PJNZ", package = "specio")
spectrum_region_code <- read_spectrum_region_code(pjnz_path)

}
