% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epp_results.R
\name{read_spu}
\alias{read_spu}
\title{Read EPP uncertainty results from SPU file.}
\usage{
read_spu(pjnz_path)
}
\arguments{
\item{pjnz_path}{Path to the PJNZ file.}
}
\value{
List containing uncertainty results for prevalence and incidence
}
\description{
Read EPP uncertainty results from SPU file.
}
