% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{tidy_data_for_epp}
\alias{tidy_data_for_epp}
\title{Tidy data for EPP.}
\usage{
tidy_data_for_epp(dat, is_percentage = FALSE, add_names = FALSE,
  site_names = NA)
}
\arguments{
\item{dat}{Data to be tidied.}

\item{is_percentage}{True if data represents a percentage.}

\item{add_names}{True if names should be added to the data.}

\item{site_names}{List of site names to be used for the data.}
}
\value{
The tidied and labelled data.
}
\description{
Helper to tidy data and add labels for use in EPP.
}
\details{
This will convert any '-1' in the data to an NA as the Java representation
of the data uses -1 to indicate no data.
If is_percentage is true it will divide the data by 100 as Java
representation of data is in percentage but EPP needs it as a probability.
If add_names is true it will add labels to the matrix or vector. If a vector
then it will use years for the names. If a matrix then it will add year
labels to the columns and site labels specified by site_names to the rows.
}
\keyword{internal}
