% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_agegr_labels}
\alias{get_agegr_labels}
\title{Get the age group labels.}
\usage{
get_agegr_labels(between_15_49_only = FALSE)
}
\arguments{
\item{between_15_49_only}{Whether to get age ranges between 15 and 49 only.}
}
\description{
Get the age group labels.
}
\keyword{internal}
