% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{get_anc_data}
\alias{get_anc_data}
\title{Extract and prepare ANC related data from ProjectionSet.}
\usage{
get_anc_data(projection_set, input_mode)
}
\arguments{
\item{projection_set}{An xml_nodeset representing a ProjectionSet java
object.}

\item{input_mode}{}
}
\description{
Helper function which handles extracting and tidying ANC data from the
xml_nodeset representing ProjectionSet java object.
}
\keyword{internal}
