% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{get_array_property_data}
\alias{get_array_property_data}
\title{Get data for a property from the xml_nodeset.}
\usage{
get_array_property_data(nodeset, property, site_names = NULL)
}
\arguments{
\item{nodeset}{The xml nodeset containing the property.}

\item{property}{The property to get data for.}

\item{site_names}{Optional list of site names to use for matrix row names.}
}
\value{
Parsed property data.
}
\description{
Will get array data from one xml_node from an xml_nodeset. If node
represents a vector then a vector is returned. If node repsents a matrix then
parse data and return matrix.
}
\keyword{internal}
