% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_fert_rat}
\alias{get_fert_rat}
\title{Get fertility ratio property.}
\usage{
get_fert_rat(tag, tag_data, ...)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{...}{Other params made available by tag configuration.}
}
\description{
Get fertility ratio property.
}
\keyword{internal}
