% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_accessors.R
\name{get_tag_data}
\alias{get_tag_data}
\title{Get a tag from the DP file.}
\usage{
get_tag_data(tag, tag_data, ...)
}
\arguments{
\item{tag}{The name of the tag getting data for.}

\item{tag_data}{Tag data to be parsed.}

\item{...}{Other params made available calling function.}
}
\value{
The tidied tag data.
}
\description{
Parses data for a particular property.
}
\keyword{internal}
