% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specio.R
\name{parse_array}
\alias{parse_array}
\title{Parse xml node containing array to vector.}
\usage{
parse_array(xml_node)
}
\arguments{
\item{xml_node}{`xml_node` object representing an array.}
}
\value{
A vector representing the array.
}
\description{
Parses XML representation of java array from EPP .xml file.
}
\keyword{internal}
