% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{parse_matrix}
\alias{parse_matrix}
\title{Parse xml node representing matrix to an R matrix.}
\usage{
parse_matrix(xml_node)
}
\arguments{
\item{xml_node}{`xml_node` object representing a matrix.}
}
\value{
Matrix
}
\description{
Parses XML representation of java array of arrays from EPP .xml file.
}
\details{
This uses parse_array to parse each array in turn then groups together into
a matrix. If an array is missing at specified index this will add an NA row
to the returned matrix at that index.
}
\keyword{internal}
