% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_file_utils.R
\name{row_contains_data}
\alias{row_contains_data}
\title{Check whether a row contains data.}
\usage{
row_contains_data(dp_data, index)
}
\arguments{
\item{dp_data}{The dp_data to check.}

\item{index}{The index of the row to check.}
}
\description{
A row is considered to contain data if the fields from the 'Data' column
to the final column are not all NA values.
}
\keyword{internal}
